feat(limits): Increase TID to 16-bit and clock to 48-bit
Technical changes:
- Epoch type: uint32 → uint64 (16-bit TID + 48-bit clock)
- MaxThreads: 256 → 65,536 goroutines
- TIDBits: 8 → 16 (supports 65K concurrent goroutines)
- ClockBits: 24 → 48 (281T operations, practically unlimited)

Impact:
- Fixes TID overflow at 256 goroutines (production issue)
- Fixes clock overflow at 16M operations (tight loops)
- Memory: Epoch 4→8 bytes, VectorClock 1KB→256KB (acceptable)
- Performance: Go 1.25+ has excellent 64-bit performance

Core packages passing:
- internal/race/epoch: all tests pass, 0 linter issues
- internal/race/vectorclock: all tests pass, 0 linter issues
- internal/race/syncshadow: all tests pass

Remaining: Type conversion fixes needed in test files for
goroutine/detector/api/shadowmem packages (follow-up task).

Refs: v0.2-task4-increase-limits

+++ internal/race/detector/detector_bench_test.go
@@ -97,8 +97,8 @@ func BenchmarkOnWrite_WithRace(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		// Set previous write to future epoch to trigger race.
 		vs.W = epoch.NewEpoch(1, 1000000)
-		ctx.C.Set(1, uint32(i))
-		ctx.Epoch = epoch.NewEpoch(1, uint32(i))
+		ctx.C.Set(1, uint64(i))
+		ctx.Epoch = epoch.NewEpoch(1, uint64(i))
 
 		d.OnWrite(addr, ctx)
 	}
@@ -344,8 +344,8 @@ func BenchmarkOnRead_WithRace(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		// Set previous write to future epoch to trigger race.
 		vs.W = epoch.NewEpoch(1, 1000000)
-		ctx.C.Set(1, uint32(i))
-		ctx.Epoch = epoch.NewEpoch(1, uint32(i))
+		ctx.C.Set(1, uint64(i))
+		ctx.Epoch = epoch.NewEpoch(1, uint64(i))
 
 		d.OnRead(addr, ctx)
 	}
